Has the makefiles needed to work with the ITU-T sample code provided for MOS calculation using PESQ method
ITU-T recommendation P.862 released in 11/2015 has the reference
implementatio for calculating MOS score.
However the code is provided without any makefile. In the process of
verifying the reference implementation i had to create the linux
makefiles and windows solution. Makefiles are added here so that anyone
can use it directly.
1. Makefile-linux: Can be directly used as it is.
Just rename the file as Makefile and keep it inside the source folder.
Run make prepare
- which will make the /bin and /obj folder.

make clean
- to clean the binary and .os if prepared earlier.

make
- to make the binary.

Binary will be named as "pesq" and can be located inside the /bin
folder.